server/app/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.11.7
sqlalchemy==2.0.41
uvicorn==0.34.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
pydantic_settings==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.27.0